contest div3 update today

+++ codeforces/contest_06042024/c.cpp
@@ -18,13 +18,13 @@ int main() {
         ll minelements = k / m;
         if (k%m!=0) minelements++;
         priority_queue<pair<ll, ll>> pq;
-        for (int i=0; i<n; i++) {
+        for (ll i=0; i<n; i++) {
             pq.push({a[i], i});
             if (pq.size() > minelements) {
                 pq.pop();
             }
         }
-        int minsum = 0;
+        ll minsum = 0;
         priority_queue<pair<ll, ll>, vector<pair<ll, ll>>, greater<pair<ll, ll>>> pqindices;
         ll maxelement = pq.top().first;
         while (!pq.empty()) {
